Fixed auth and added test route.
