+++ server.js
@@ -9,53 +9,66 @@ var port = process.env.PORT || 3000;
 // Including the configured express and mongoose objects
 var mongoose = require('./server/config/mongoose');
 
-// EXPRESS CONFIGURATION FILE
 
-// node libraries and configuration file
-var   express = require('express')
-    , cors = require('cors')
-    , bodyParser = require('body-parser')
-	, passport = require('passport');
-
-    var app = express();
-
-    // MIDDLEWARE NEEDED BOTH FOR DEV AND PRODUCTION
-
-    app.use(cors());
-    app.use(bodyParser.json());
-    app.use(bodyParser.urlencoded(
-        {
-            extended: true
-        }));
-
-	var passport = require('passport');
-	app.use(passport.initialize());
-	
-	
-    // HERE WE CONFIGURE THE ROUTES
-    require('./server/friends/friend.server.routes')(app);
-	
-	app.get('/auth/familysearch',
-		passport.authenticate('familysearch'));
+var express = require('express')
+	, app = express()
+	, bodyParser = require('body-parser')
+	, cors = require('cors')
+	, port = 3000
+	, passport = require("passport")
+	, db = mongoose();
 
 
-	app.get('/auth/familysearch/callback',
-		passport.authenticate('familysearch', { failureRedirect: '/' }),
-		function (req, res) {
-			res.redirect('http://localhost:3000/');
+app.use(bodyParser.json());
+app.use(cors());
+app.use(express.static(__dirname + '/public'));
 
-		});
+app.use(passport.initialize())
 
-	app.get('/logout', function (req, res) {
-		req.logout();
-		res.redirect('/');
+passport.serializeUser(function (user, done) {
+	done(null, user);
+});
+
+passport.deserializeUser(function (obj, done) {
+	done(null, obj);
+});
+
+
+var FamilySearchStrategy = require('passport-familysearch').Strategy;
+
+passport.use(new FamilySearchStrategy({
+	authorizationURL: 'https://sandbox.familysearch.org/cis-web/oauth2/v3/authorization',
+    tokenURL: 'https://sandbox.familysearch.org/cis-web/oauth2/v3/token',
+    devKey: "a02j000000BpvVCAAZ",
+    callbackURL: "http://localhost:3000/auth/familysearch/callback",
+	environment: 'sandbox'
+},
+	function (accessToken, refreshToken, profile, done) {
+		console.log("access token:", accessToken);
+		console.log("profile", profile);
+		return done(null, profile);
+	}
+	));
+
+
+app.get('/auth/familysearch',
+	passport.authenticate('familysearch'));
+
+
+app.get('/auth/familysearch/callback',
+	passport.authenticate('familysearch', { failureRedirect: '/' }),
+	function (req, res) {
+		res.redirect('http://localhost:3000/');
+		
 	});
 
+app.get('/logout', function (req, res) {
+	req.logout();
+	res.redirect('/');
+});
 
 
-// Let's run this!
-var db = mongoose(),
-    app = express();
+
 
 app.use(express.static('./public/'));
